Feat/improved input (#1)
* Added Seconds Spinner

Implemented a seconds spinner. limited to 60 seconds because we'll be adding minute and hour spinners that autoincrement based on seconds incrementing next.

* Hours, Minutes input added

The minutes and hours spinners have been added, and auto-increment one-another appropriately. Now comes the fun part: ui design

* UI Redesign

Some updates to the UI portion of the application. Next up: Implement decrements in a simililar fashion to the increments.

* Decrement Update

The time "dials" now have behavior to handle both decrements and increments to the respective controls.